+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Support for connecting to unix sockets. Thanks to [@aschey](https://github.com/aschey)
+- Support google.protobuf.Struct. Thanks to [@n0trace](https://github.com/n0trace)
 
 ## [v0.5.0] - 2021-04-26